backend/requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0

# LangChain and AI
langchain>=1.2.0
langchain-chroma>=1.1.0
langchain-community>=0.4.1
langchain-huggingface>=1.2.0
langchain-ollama>=1.0.1
langchain-text-splitters>=0.3.0

# Vector DB and Embeddings
chromadb>=0.4.0
sentence-transformers>=5.2.0

# PDF Processing
pymupdf>=1.26.7

# Utilities
python-dotenv>=1.0.0
